src/requirements.txt
faiss-cpu==1.8.0.post1
langchain==0.2.5
langchain-cli==0.0.25
langchain-community==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.10
langchain-text-splitters==0.2.1
langchainhub==0.1.20
openai==1.35.3
pypdf==4.2.0
python-dotenv==1.0.1
